move selectors to features
